reorganise directory. simplify. rerun tests
