+++ packages/document-issue-io/src/document_issue_io/media/mf_medium.png
