+++ packages/document-issue-io/src/document_issue_io/media/titleblock_cambridge.png
